+++ src/operations/Completion.cpp
@@ -3,6 +3,19 @@
 #include "LSP/Protocol.hpp"
 #include "LSP/LuauExt.hpp"
 
+/// Defining sort text levels assigned to completion items
+/// Note that sort text is lexicographically
+namespace SortText
+{
+static constexpr const char* TableProperties = "0";
+static constexpr const char* CorrectTypeKind = "1";
+static constexpr const char* CorrectFunctionResult = "2";
+static constexpr const char* Default = "3";
+static constexpr const char* WrongIndexType = "4";
+static constexpr const char* AutoImports = "5";
+static constexpr const char* Keywords = "6";
+} // namespace SortText
+
 static std::optional<Luau::AutocompleteEntryMap> nullCallback(std::string tag, std::optional<const Luau::ClassTypeVar*> ptr)
 {
     return std::nullopt;
@@ -264,6 +277,7 @@ void WorkspaceFolder::suggestImports(
             item.detail = "Auto-import";
             item.documentation = {lsp::MarkupKind::Markdown, codeBlock("lua", importText)};
             item.insertText = service;
+            item.sortText = SortText::AutoImports;
 
             lsp::Position placement{lineNumber, 0};
             item.additionalTextEdits.emplace_back(lsp::TextEdit{{placement, placement}, importText});
@@ -297,9 +311,21 @@ std::vector<lsp::CompletionItem> WorkspaceFolder::completion(const lsp::Completi
         lsp::CompletionItem item;
         item.label = name;
         item.deprecated = entry.deprecated;
+        item.sortText = SortText::Default;
 
         if (entry.documentationSymbol)
             item.documentation = {lsp::MarkupKind::Markdown, printDocumentation(client->documentation, *entry.documentationSymbol)};
+
+        if (entry.typeCorrect == Luau::TypeCorrectKind::Correct)
+            item.sortText = SortText::CorrectTypeKind;
+        else if (entry.typeCorrect == Luau::TypeCorrectKind::CorrectFunctionResult)
+            item.sortText = SortText::CorrectFunctionResult;
+        else if (entry.wrongIndexType)
+            item.sortText = SortText::WrongIndexType;
+        else if (entry.kind == Luau::AutocompleteEntryKind::Property)
+            item.sortText = SortText::TableProperties;
+        else if (entry.kind == Luau::AutocompleteEntryKind::Keyword)
+            item.sortText = SortText::Keywords;
 
         switch (entry.kind)
         {