+++ src/operations/SemanticTokens.cpp
@@ -4,6 +4,26 @@
 #include "LSP/LuauExt.hpp"
 #include "LSP/SemanticTokens.hpp"
 
+static void fillBuiltinGlobals(std::unordered_map<Luau::AstName, Luau::TypeId>& builtins, const Luau::AstNameTable& names, const Luau::ScopePtr& env)
+{
+    Luau::ScopePtr current = env;
+    while (true)
+    {
+        for (auto& [global, binding] : current->bindings)
+        {
+            Luau::AstName name = names.get(global.c_str());
+
+            if (name.value)
+                builtins.insert_or_assign(name, binding.typeId);
+        }
+
+        if (current->parent)
+            current = current->parent;
+        else
+            break;
+    }
+}
+
 enum struct AstLocalInfo
 {
     // local is self
@@ -19,9 +39,12 @@ static lsp::SemanticTokenTypes inferTokenType(Luau::TypeId* ty, lsp::SemanticTok
 
     auto followedTy = Luau::follow(*ty);
 
-    if (Luau::get<Luau::FunctionTypeVar>(followedTy))
+    if (auto ftv = Luau::get<Luau::FunctionTypeVar>(followedTy))
     {
-        return lsp::SemanticTokenTypes::Function;
+        if (ftv->hasSelf)
+            return lsp::SemanticTokenTypes::Method;
+        else
+            return lsp::SemanticTokenTypes::Function;
     }
     else if (Luau::get<Luau::IntersectionTypeVar>(followedTy))
     {
@@ -46,11 +69,41 @@ struct SemanticTokensVisitor : public Luau::AstVisitor
     Luau::ModulePtr module;
     std::vector<SemanticToken> tokens;
     std::unordered_map<Luau::AstLocal*, AstLocalInfo> localMap;
+    std::unordered_map<Luau::AstName, Luau::TypeId> builtinGlobals;
+    std::unordered_set<Luau::AstType*> syntheticTypes;
 
-    explicit SemanticTokensVisitor(Luau::ModulePtr module)
+    explicit SemanticTokensVisitor(Luau::ModulePtr module, std::unordered_map<Luau::AstName, Luau::TypeId> builtinGlobals)
         : module(module)
+        , builtinGlobals(builtinGlobals)
     {
     }
+
+    // HACK: Luau introduces some synthetic tokens in the AST for types
+    // { T } gets converted to { [number]: T } (where number is an introduced AstTypeReference)
+    // string? gets converted to string | nil (where nil is an introduced AstTypeReference)
+    // We do not want to highlight these synthetic tokens, as they don't exist in real code.
+    // We apply visitors in these locations to try and check for synthetic tokens, and mark them as such
+    bool visit(Luau::AstTypeTable* table) override
+    {
+        // If the indexer location is the same as a result type, the indexer was synthetically added
+        if (table->indexer && table->indexer->indexType->location == table->indexer->resultType->location)
+            syntheticTypes.emplace(table->indexer->indexType);
+        return true;
+    }
+
+    bool visit(Luau::AstTypeUnion* unionType) override
+    {
+        // If the union contains a "nil" part, but it has the the size of only 1 column, then it is synthetic
+        // Note that the union could contain > 2 parts:
+        //  T? -> T | nil
+        //  U | V? -> U | V | nil
+        for (const auto& ty : unionType->types)
+            if (auto ref = ty->as<Luau::AstTypeReference>())
+                if (!ref->prefix && !ref->hasParameterList && ref->name == "nil" && ref->location.end.column == ref->location.begin.column + 1)
+                    syntheticTypes.emplace(ref);
+        return true;
+    }
+
 
     bool visit(Luau::AstType* type) override
     {
@@ -59,6 +112,10 @@ struct SemanticTokensVisitor : public Luau::AstVisitor
 
     bool visit(Luau::AstTypeReference* ref) override
     {
+        // Ignore synthetic type references
+        if (syntheticTypes.find(ref) != syntheticTypes.end())
+            return false;
+
         // HACK: The location information provided only gives location for the whole reference
         // but we do not want to highlight punctuation inside of it (Module.Type<Param> should not highlight . < >)
         // So we use the start position and consider the end positions separately.
@@ -170,12 +227,46 @@ struct SemanticTokensVisitor : public Luau::AstVisitor
         return true;
     }
 
-    // bool visit(Luau::AstExprGlobal* local) override
-    // {
-    //     auto type = inferTokenType(module->astTypes.find(local), lsp::SemanticTokenTypes::Namespace);
-    //     tokens.emplace_back(SemanticToken{local->location.begin, local->location.end, type, lsp::SemanticTokenModifiers::None});
-    //     return true;
-    // }
+    bool visit(Luau::AstExprGlobal* global) override
+    {
+        auto it = builtinGlobals.find(global->name);
+        if (it != builtinGlobals.end() && strlen(global->name.value) > 0)
+        {
+            // SPECIAL CASE: if name is "Enum", classify it as an enum
+            if (global->name == "Enum")
+            {
+                tokens.emplace_back(SemanticToken{
+                    global->location.begin, global->location.end, lsp::SemanticTokenTypes::Enum, lsp::SemanticTokenModifiers::DefaultLibrary});
+            }
+            // If it starts with an uppercase letter, flag it as a class
+            // Otherwise, flag it as a builtin
+            else if (isupper(global->name.value[0]))
+            {
+                tokens.emplace_back(SemanticToken{
+                    global->location.begin, global->location.end, lsp::SemanticTokenTypes::Class, lsp::SemanticTokenModifiers::DefaultLibrary});
+            }
+            else
+            {
+                auto type = inferTokenType(&it->second, lsp::SemanticTokenTypes::Variable);
+                tokens.emplace_back(SemanticToken{global->location.begin, global->location.end, type,
+                    lsp::SemanticTokenModifiers::DefaultLibrary | lsp::SemanticTokenModifiers::Readonly});
+            }
+        }
+        else
+        {
+            auto ty = module->astTypes.find(global);
+            if (!ty)
+                return true;
+
+            auto type = inferTokenType(ty, lsp::SemanticTokenTypes::Variable);
+            if (type == lsp::SemanticTokenTypes::Variable)
+                return true;
+
+            tokens.emplace_back(SemanticToken{global->location.begin, global->location.end, type, lsp::SemanticTokenModifiers::None});
+        }
+
+        return true;
+    }
 
     bool visit(Luau::AstExprIndexName* index) override
     {
@@ -183,12 +274,31 @@ struct SemanticTokensVisitor : public Luau::AstVisitor
         if (!parentTy)
             return true;
 
+        auto parentIsBuiltin = false;
+        auto parentIsEnum = false;
+        if (auto global = index->expr->as<Luau::AstExprGlobal>())
+        {
+            parentIsBuiltin = builtinGlobals.find(global->name) != builtinGlobals.end();
+            if (parentIsBuiltin && global->name == "Enum")
+                parentIsEnum = true;
+        }
+
         auto ty = Luau::follow(*parentTy);
         if (auto prop = lookupProp(ty, std::string(index->index.value)))
         {
-            auto type = inferTokenType(&prop->type, lsp::SemanticTokenTypes::Property);
+            auto defaultType = lsp::SemanticTokenTypes::Property;
+            if (parentIsEnum)
+                defaultType = lsp::SemanticTokenTypes::Enum;
+            else if (Luau::hasTag(prop->tags, "EnumItem"))
+                defaultType = lsp::SemanticTokenTypes::EnumMember;
+
+            auto type = inferTokenType(&prop->type, defaultType);
             auto modifiers = lsp::SemanticTokenModifiers::None;
-            if (types::isMetamethod(std::string(index->index.value)))
+            if (parentIsBuiltin)
+            {
+                modifiers = modifiers | lsp::SemanticTokenModifiers::DefaultLibrary | lsp::SemanticTokenModifiers::Readonly;
+            }
+            else if (types::isMetamethod(std::string(index->index.value)))
             {
                 modifiers = modifiers | lsp::SemanticTokenModifiers::DefaultLibrary;
             }
@@ -223,9 +333,12 @@ struct SemanticTokensVisitor : public Luau::AstVisitor
     }
 };
 
-std::vector<SemanticToken> getSemanticTokens(Luau::ModulePtr module, Luau::SourceModule* sourceModule)
+std::vector<SemanticToken> getSemanticTokens(const Luau::Frontend& frontend, const Luau::ModulePtr module, const Luau::SourceModule* sourceModule)
 {
-    SemanticTokensVisitor visitor{module};
+    std::unordered_map<Luau::AstName, Luau::TypeId> builtinGlobals;
+    fillBuiltinGlobals(builtinGlobals, *sourceModule->names, frontend.typeChecker.globalScope);
+
+    SemanticTokensVisitor visitor{module, builtinGlobals};
     visitor.visit(sourceModule->root);
     return visitor.tokens;
 }
@@ -283,7 +396,7 @@ std::optional<lsp::SemanticTokens> WorkspaceFolder::semanticTokens(const lsp::Se
     if (!sourceModule || !module)
         return std::nullopt;
 
-    auto tokens = getSemanticTokens(module, sourceModule);
+    auto tokens = getSemanticTokens(frontend, module, sourceModule);
     lsp::SemanticTokens result;
     result.data = packTokens(tokens);
     return result;